Add OpenStack testing (via ssh) to travis ci
We need to be able to run CI against OpenStack. This approach uses ssh
to run the testing commands from a jump host maintained by Cisco Cloud.
It requires the secret env vars OS_IP and OS_KEY, which are the IP
address of the jump box, and an ssh private key, respectively. It also
requires the creation of a travis user on the jump host.

Most of the changes here are if/thens for whether we need to test
against OpenStack or not. My hope is that when we switch our deployments
to a container, this will simplify things in this file also. Right now,
it feels a bit hacky.

As part of these changes, I altered the semantics of the cleanup script:
instead of calling terraform in two lines, we will use the builtin
`travis_retry` command. It will attempt the destroy 3 times, which
should be more than enough retries for the terraform AWS bug.